acpi: refactor scope lookup code and move it into the entity pkg
